+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+Added day of the week above the date.
+
+Preparations to support the Pebble Time Round.
+
 ## [1.1] - 2015-06-14
 
 Added battery status and Bluetooth connection indicators.